+++ test/libsolidity/syntaxTests/userDefinedValueType/delete.sol
@@ -1,0 +1,5 @@
+type MyInt is int;
+function f() pure {
+    MyInt i = MyInt.wrap(1);
+    delete i;
+}